requirements.txt
yoyo-migrations>=6.0,<6.1